+++ R/cache-helpers.R
@@ -388,17 +388,22 @@ getFunctionName <- function(FUN, originalDots, ..., overrideCall, isPipe) { # no
       callIndices <- .grepSysCalls(scalls, pattern = paste0("^", overrideCall))
       functionCall <- scalls[callIndices]
     } else {
-      callIndices <- .grepSysCalls(scalls, pattern = "^Cache|^SpaDES::Cache|^reproducible::Cache")
+      callIndices <- .grepSysCalls(scalls,
+        pattern = "^Cache|^SpaDES::Cache|^reproducible::Cache|^cloudCache")
+      callIndicesDoCall <- .grepSysCalls(scalls, pattern = "^do.call")
+      doCall1st2Elements <- lapply(scalls[callIndicesDoCall], function(x) x[1:2])
+      callIndicesDoCall <- callIndicesDoCall[grep("Cache", doCall1st2Elements)]
       # The next line takes too long to grep if scalls has enormous objects
       # callIndices <- grep(scalls, pattern = "^Cache|^SpaDES::Cache|^reproducible::Cache")
+      callIndices <- unique(sort(c(callIndices, callIndicesDoCall)))
       functionCall <- scalls[callIndices]
     }
     if (length(functionCall)) {
       # for() loop is a work around for R-devel that produces a different final call in the
       # sys.calls() stack which is NOT .Method ... and produces a Cache(FUN = FUN...)
-      for (callIndex in rev(callIndices)) {
+      for (callIndex in head(rev(callIndices), 2)) {
         if (!missing(overrideCall)) {
-          env <- sys.frames()[[callIndices]]
+          env <- sys.frames()[[callIndex]]
           matchedCall <- match.call(get(overrideCall, envir = env), scalls[[callIndex]])#parse(text = callIndex))
           forms <- tryCatch("FUN" %in% formalArgs(overrideCall), error = function(x) NULL)
           if (!is.null(forms)) {
@@ -407,11 +412,36 @@ getFunctionName <- function(FUN, originalDots, ..., overrideCall, isPipe) { # no
             functionName <- matchedCall[[2]]
           }
         } else {
-          matchedCall <- match.call(Cache, scalls[[callIndex]])#parse(text = callIndex))
-          functionName <- matchedCall$FUN
+          foundCall <- FALSE
+          if (exists("callIndicesDoCall", inherits = FALSE))
+            if (length(callIndicesDoCall) > 0) {
+              if (callIndex %in% callIndicesDoCall) {
+                mcDoCall <- match.call(do.call, scalls[[callIndex]])
+                for (i in 1:2) {
+                  fnLookup <- try(eval(mcDoCall$args, envir = sys.frames()[[callIndex - i]]), silent = TRUE)
+                  if (!is(fnLookup, "try-error"))
+                    break
+                }
+                functionName <- if (isTRUE("FUN" %in% names(fnLookup)))
+                  fnLookup$FUN
+                else
+                  fnLookup[[1]]
+
+                foundCall <- TRUE
+              }
+            }
+          if (!foundCall) {
+            matchedCall <- match.call(Cache, scalls[[callIndex]])#parse(text = callIndex))
+            functionName <- matchedCall$FUN
+          }
         }
-        functionName <- deparse(functionName, width.cutoff = 300)
-        if (all(functionName != c("FUN"))) break
+        functionName <- if (is(functionName, "name")) {
+          deparse(functionName, width.cutoff = 300)
+        } else {
+          "FUN"
+        }
+
+        if (all(functionName != c("FUN")) && all(functionName != c("NULL"))) break
       }
     } else {
       functionName <- ""
@@ -865,6 +895,7 @@ copyFile <- Vectorize(copySingleFile, vectorize.args = c("from", "to"))
 #' @rdname cacheHelper
 #' @importFrom methods slotNames
 #' @importFrom digest digest
+#' @importFrom fastdigest fastdigest
 #' @importFrom raster res crs extent
 .digestRasterLayer <- function(object, length, algo, quick) {
   # metadata -- only a few items of the long list because one thing (I don't recall)
@@ -873,9 +904,14 @@ copyFile <- Vectorize(copySingleFile, vectorize.args = c("from", "to"))
   sn <- sn[!(sn %in% c("min", "max", "haveminmax", "names", "isfactor",
                        "dropped", "nlayers", "fromdisk", "inmemory", "offset", "gain"))]
   dataSlotsToDigest <- lapply(sn, function(s) slot(object@data, s))
-  dig <- digest(append(list(dim(object), res(object), crs(object),
+  if (isTRUE(getOption("reproducible.useNewDigestAlgorithm")))
+    dig <- digest(append(list(dim(object), res(object), crs(object),
                          extent(object)), dataSlotsToDigest),
                 algo = algo) # don't include object@data -- these are volatile
+  else
+    dig <- fastdigest(append(list(dim(object), res(object), crs(object),
+                              extent(object)), dataSlotsToDigest)) # don't include object@data -- these are volatile
+
   if (nzchar(object@file@name)) {
     # if the Raster is on disk, has the first length characters;
     filename <- if (isTRUE(endsWith(basename(object@file@name), suffix = ".grd"))) {
@@ -887,7 +923,12 @@ copyFile <- Vectorize(copySingleFile, vectorize.args = c("from", "to"))
     dig2 <- .robustDigest(asPath(filename, 2), length = length, quick = quick, algo = algo)
     dig <- c(dig, dig2)
   }
-  dig <- digest(dig, algo = algo)
+
+  if (isTRUE(getOption("reproducible.useNewDigestAlgorithm")))
+    dig <- digest(dig, algo = algo)
+  else
+    dig <- fastdigest(dig)
+  dig
 }
 
 #' Recursive copying of nested environments, and other "hard to copy" objects