+++ NEWS.md
@@ -9,9 +9,11 @@ version 0.2.5.9000
 
 ## New features
 
+* remove `%>%` pipe that was long ago deprecated. User should use `%C%` if they want a pipe that is Cache-aware. See examples.
+* Full rewrite of all `options` descriptions now in `reproducible`, see `?reproducibleOptions`
 * now `cacheRepo` and `options("reproducible.cachePath")` can take a vector of paths. Similar to how .libPaths() works for libraries, `Cache` will search first in the first entry in the `cacheRepo`, then the second etc. until it finds an entry. It will only write to the first entry.
 * new value for the option: `options("reproducible.useCache" = "devMode")`. The point of this mode is to facilitate using the Cache when functions and datasets are continually in flux, and old Cache entries are likely stale very often. In `devMode`, the cache mechanism will work as normal if the Cache call is the first time for a function OR if it successfully finds a copy in the cache based on the normal Cache mechanism. It *differs* from the normal Cache if the Cache call does *not* find a copy in the `cacheRepo`, but it does find an entry that matches based on `userTags`. In this case, it will delete the old entry in the `cacheRepo` (identified based on matching `userTags`), then continue with normal `Cache`. For this to work correctly, `userTags` must be unique for each function call. This should be used with caution as it is still experimental.
-* slight tweak to how digests are calculated. This will cause some existing caches to not work correctly. To allow for a transition period, the "new" algorigthm is not default, but can be set with `options("reproducible.useNewDigestAlgorithm" = TRUE)`. There is a warning of this upcoming change.
+* change to how hashes are calculated. This will cause existing caches to not work correctly. To allow for a transition period, the "new" algorigthm is default, but can be unset with `options("reproducible.useNewDigestAlgorithm" = FALSE)`. There is a message of this change on package load.
 * add experimental `cloud*` functions, especially `cloudCache` which allows sharing of Cache among collaborators
 * updated `assessDataType` to consolidate `assessDataTypeGDAL` and `assessDataType` into single function (#71, @ianmseddy)
 * `cc`: new function -- a shortcut for some commonly used options for `clearCache()`